make it work with elementor pages
`post_content` isn't always populated at this point. Elementor for instance uses the `the_content` hook, I think also normal posts as well. That's why this way it would be more compatible. Would be cool if this could be incorporated within the next release, so we don't have to monkey patch it every time :) 

thanks